Prototipo misil funcionando.
