+++ src/core/domain/types/base/index.ts
@@ -1,0 +1,3 @@
+export { IImage } from "./IImage";
+export { IShape } from "./IShape";
+export { IVisual } from "./IVisual";
