+++ src/core/domain/types/index.ts
@@ -4,3 +4,4 @@ export { Vector3 } from "./Vector3";
 export { Acceleration } from "./Acceleration";
 export { Transform } from "./Transform";
 export { Renderable } from "./Renderable";
+export * from "./base";
